+++ source/geh_common/release-notes.md
@@ -1,8 +1,16 @@
 # GEH Common Release Notes
 
-## Version 5.8.8
+## Version 5.8.10
 
 - Adding functionally to execute_statement, such that it can handle both sync and async calls.
+
+## Version 5.8.9
+
+Add `geh_common.pyspark.clamp` functions to clamp periods.
+
+## Version 5.8.8
+
+- Removed set notation characters from nsg6 consumption contract
 
 ## Version 5.8.7